// src/types/supabase.ts
export type Task = {
  id: string;
  created_at: string;
  title: string;
  description: string;
  due_date: string;
  status: 'pending' | 'completed';
  priority: 'low' | 'medium' | 'high';
  related_to_type: 'lead' | 'contact';
  related_to_id: string;
  assigned_to: string;
  organization_id: string;
  branch_id: string;
};